# Expand grid concepts to n dimensions