--- a/beluga/include/beluga/sensor/data/dense_grid.hpp
+++ b/beluga/include/beluga/sensor/data/dense_grid.hpp
@@ -21,6 +21,7 @@
 #include <beluga/sensor/data/regular_grid.hpp>
 
 #include <Eigen/Core>
+#include <beluga/eigen_compatibility.hpp>
 
 /**
  * \file