--- a/beluga/include/beluga/sensor/data/sparse_value_grid.hpp
+++ b/beluga/include/beluga/sensor/data/sparse_value_grid.hpp
@@ -27,21 +27,22 @@
 
 namespace beluga {
 
-/// Generic 2D sparse value regular grid.
+/// Generic N dimensional sparse value regular grid.
 /**
  * \tparam MapType:
- *    Associative container representing a mapping from a Eigen::Vector2i to a value of type MapType::mapped_type.
+ *    Associative container representing a mapping from a Eigen::Vector<int, NDim> to a value of type MapType::mapped_type.
  *    It should implement a subset of standard library's associative containers public API.
  *    In particular, given 'm' a possibly const instance of MapType:
- *         - MapType::key_type must be Eigen::Vector2i.
+ *         - MapType::key_type must be Eigen::Vector<int, NDim>.
  *         - MapType::mapped_type should be the value type of the associative container entries.
- *         - 'm.at(const Eigen::Vector2i& cell_index) const' should return a const reference to 'MapType::value_type'
+ *         - 'm.at(const Eigen::Vector<int, NDim>& cell_index) const' should return a const reference to 'MapType::value_type'
  * representing the value at that index, or throw 'std::out_of_range' if it doesn't exist.
- *        - 'm.find(const Eigen::Vector2i&)' should follow the same API as
+ *        - 'm.find(const Eigen::Vector<int, NDim>&)' should follow the same API as
  * [std::map](https://en.cppreference.com/w/cpp/container/map/find).
+ * \tparam NDim: Dimension of the grid.
  */
-template <typename MapType>
-class SparseValueGrid : public BaseRegularGrid2<SparseValueGrid<MapType>> {
+template <typename MapType, int NDim>
+class SparseValueGrid : public BaseRegularGrid<SparseValueGrid<MapType, NDim>, NDim> {
  public:
   /// Construct without data and with a 1 cell/meter resolution.
   SparseValueGrid() = default;
@@ -51,7 +52,8 @@ class SparseValueGrid : public BaseRegularGrid2<SparseValueGrid<MapType>> {
   using mapped_type = typename map_type::mapped_type;
   /// Key type for the key-value pairs.
   using key_type = typename map_type::key_type;
-  static_assert(std::is_same_v<key_type, Eigen::Vector2i>);
+
+  static_assert(std::is_same_v<key_type, Eigen::Vector<int, NDim>>);
   /// Constructs the grid.
   /**
    * \param data Grid data.
@@ -71,8 +73,8 @@ class SparseValueGrid : public BaseRegularGrid2<SparseValueGrid<MapType>> {
   [[nodiscard]] const map_type& data() const { return data_; }
 
   /// Gets grid data at cell_index or std::nullopt if it's not present.
-  [[nodiscard]] std::optional<mapped_type> data_at(const Eigen::Vector2i& cell_index) const {
-    auto itr = data_.find(cell_index);
+  [[nodiscard]] std::optional<mapped_type> data_at(const Eigen::Vector<int, NDim>& cell_index) const {
+    const auto itr = data_.find(cell_index);
     if (itr == data_.end()) {
       return std::nullopt;
     }
@@ -80,13 +82,8 @@ class SparseValueGrid : public BaseRegularGrid2<SparseValueGrid<MapType>> {
   }
 
   /// Gets grid data at real coordinates 'coordinates' or std::nullopt if it's not present.
-  [[nodiscard]] std::optional<mapped_type> data_near(const Eigen::Vector2d& coordinates) const {
+  [[nodiscard]] std::optional<mapped_type> data_near(const Eigen::Vector<double, NDim>& coordinates) const {
     return data_at(this->self().cell_near(coordinates));
-  }
-
-  /// Gets grid data at real coordinates '{x, y}' or std::nullopt if it's not present.
-  [[nodiscard]] std::optional<mapped_type> data_near(const double x, const double y) const {
-    return data_at(this->self().cell_near(x, y));
   }
 
  private:
@@ -94,6 +91,8 @@ class SparseValueGrid : public BaseRegularGrid2<SparseValueGrid<MapType>> {
   double resolution_ = 1.0;
 };
 
+template <typename MapType>
+using SparseValueGrid2 = SparseValueGrid<MapType, 2>;
 }  // namespace beluga
 
 #endif