--- a/beluga/include/beluga/sensor/data/sparse_value_grid.hpp
+++ b/beluga/include/beluga/sensor/data/sparse_value_grid.hpp
@@ -30,16 +30,16 @@ namespace beluga {
 
 /// Generic N dimensional sparse value regular grid.
 /**
- * \tparam MapType:
- *    Associative container representing a mapping from a Eigen::Vector<int, NDim> to a value of type
- * MapType::mapped_type. It should implement a subset of standard library's associative containers public API. In
- * particular, given 'm' a possibly const instance of MapType:
- *         - MapType::key_type must be Eigen::Vector<int, NDim>.
- *         - MapType::mapped_type should be the value type of the associative container entries.
- *         - 'm.at(const Eigen::Vector<int, NDim>& cell_index) const' should return a const reference to
- * 'MapType::value_type' representing the value at that index, or throw 'std::out_of_range' if it doesn't exist.
- *        - 'm.find(const Eigen::Vector<int, NDim>&)' should follow the same API as
- * [std::map](https://en.cppreference.com/w/cpp/container/map/find).
+ * \tparam MapType: Associative container representing a mapping from a Eigen::Vector<int, NDim> to a value of type
+ * MapType::mapped_type. It should implement a subset of standard library's associative containers public API.
+ * In particular, given 'm' a possibly const instance of MapType:
+ * - MapType::key_type must be Eigen::Vector<int, NDim>.
+ * - MapType::mapped_type should be the value type of the associative container entries.
+ * - 'm.at(const Eigen::Vector<int, NDim>& cell_index) const' should return a const reference to
+ *   'MapType::value_type' representing the value at that index,
+ *   or throw 'std::out_of_range' if it doesn't exist.
+ * - 'm.find(const Eigen::Vector<int, NDim>&)' should follow the same API as
+ *   [std::map](https://en.cppreference.com/w/cpp/container/map/find).
  * \tparam NDim: Dimension of the grid.
  */
 template <typename MapType, int NDim>