--- a/beluga/include/beluga/sensor/ndt_sensor_model.hpp
+++ b/beluga/include/beluga/sensor/ndt_sensor_model.hpp
@@ -160,9 +160,9 @@ using NDTModelParam3d = NDTModelParam<3>;
 template <typename SparseGridT>
 class NDTSensorModel {
  public:
-  // static_assert(std::is_same_v<SparseValueGrid<typename SparseGridT::map_type>, SparseGridT>);
   /// NDT Cell type.
   using ndt_cell_type = typename SparseGridT::mapped_type;
+  static_assert(std::is_same_v<SparseValueGrid<typename SparseGridT::map_type, ndt_cell_type::num_dim>, SparseGridT>);
   static_assert(ndt_cell_type::num_dim == 2, "NDT sensor model is only implemented for 2D problems.");
   /// State type of a particle.
   using state_type = std::conditional_t<