--- a/beluga/test/beluga/sensor/data/test_regular_grid.cpp
+++ b/beluga/test/beluga/sensor/data/test_regular_grid.cpp
@@ -27,11 +27,13 @@ class TestRegularGrid2 : public beluga::BaseRegularGrid2<TestRegularGrid2> {
   [[nodiscard]] static double resolution() { return 1.; }
 };
 
+class TestRegularGrid3 : public beluga::BaseRegularGrid3<TestRegularGrid3> {
+ public:
+  [[nodiscard]] static double resolution() { return 1.; }
+};
+
 TEST(RegularGrid2, NearestCells) {
   const auto grid = TestRegularGrid2{};
-
-  EXPECT_EQ(grid.cell_near(0, 0), Eigen::Vector2i(0, 0));
-  EXPECT_EQ(grid.cell_near(1.5, 0.5), Eigen::Vector2i(1, 0));
 
   EXPECT_EQ(grid.cell_near(Eigen::Vector2d(0, 0)), Eigen::Vector2i(0, 0));
   EXPECT_EQ(grid.cell_near(Eigen::Vector2d(1.5, 0.5)), Eigen::Vector2i(1, 0));
@@ -40,8 +42,8 @@ TEST(RegularGrid2, NearestCells) {
 TEST(RegularGrid2, CoordinatesAtCell) {
   const auto grid = TestRegularGrid2{};
 
-  EXPECT_TRUE(grid.coordinates_at(0, 0).isApprox(Eigen::Vector2d(0.5, 0.5)));
-  EXPECT_TRUE(grid.coordinates_at(1, 0).isApprox(Eigen::Vector2d(1.5, 0.5)));
+  EXPECT_TRUE(grid.coordinates_at(Eigen::Vector2i{0, 0}).isApprox(Eigen::Vector2d(0.5, 0.5)));
+  EXPECT_TRUE(grid.coordinates_at(Eigen::Vector2i{1, 0}).isApprox(Eigen::Vector2d(1.5, 0.5)));
 
   EXPECT_TRUE(grid.coordinates_at(Eigen::Vector2i(0, 0)).isApprox(Eigen::Vector2d(0.5, 0.5)));
   EXPECT_TRUE(grid.coordinates_at(Eigen::Vector2i(1, 0)).isApprox(Eigen::Vector2d(1.5, 0.5)));
@@ -55,6 +57,16 @@ TEST(RegularGrid2, CoordinatesForCells) {
   EXPECT_EQ(coordinates.size(), 2);
   EXPECT_TRUE(coordinates[0].isApprox(Eigen::Vector2d(1.5, 0.5)));
   EXPECT_TRUE(coordinates[1].isApprox(Eigen::Vector2d(4.5, 1.5)));
+}
+
+TEST(RegularGrid3, CoordinatesForCells) {
+  const auto grid = TestRegularGrid3{};
+
+  const auto cells = std::vector{Eigen::Vector3i(1, 0, 0), Eigen::Vector3i(4, 1,-1)};
+  const auto coordinates = grid.coordinates_for(cells) | ranges::to<std::vector>;
+  EXPECT_EQ(coordinates.size(), 2);
+  EXPECT_TRUE(coordinates[0].isApprox(Eigen::Vector3d(1.5, 0.5, 0.5)));
+  EXPECT_TRUE(coordinates[1].isApprox(Eigen::Vector3d(4.5, 1.5, -0.5)));
 }
 
 }  // namespace