@serraramiro1 2024-06-05
@hidmic ready for a first pass. Still missing the benchmarking we talked about for the `std::floor` operation